optimize performance by streaming
